add service port project
- move core port to service port
- move reusable spec extensions to do.archictecture
