+++ Do.sln
@@ -23,6 +23,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "core", "core", "{BE392BCA-B
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Do.Test.Architecture", "test\core\Do.Test.Architecture\Do.Test.Architecture.csproj", "{45D92BDD-5EA8-4B6C-A6F0-38C9DE139A32}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Do.Blueprints.Service", "src\blueprints\Do.Blueprints.Service\Do.Blueprints.Service.csproj", "{68CBB995-C501-41BB-93B4-01A0CC618A89}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -48,6 +50,10 @@ Global
 		{45D92BDD-5EA8-4B6C-A6F0-38C9DE139A32}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{45D92BDD-5EA8-4B6C-A6F0-38C9DE139A32}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{45D92BDD-5EA8-4B6C-A6F0-38C9DE139A32}.Release|Any CPU.Build.0 = Release|Any CPU
+		{68CBB995-C501-41BB-93B4-01A0CC618A89}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{68CBB995-C501-41BB-93B4-01A0CC618A89}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{68CBB995-C501-41BB-93B4-01A0CC618A89}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{68CBB995-C501-41BB-93B4-01A0CC618A89}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{CDC3A00A-52BD-4CDA-B7D7-1FCFC5DD1CEE} = {2F297832-93E6-4750-9141-818873A769F6}
@@ -58,5 +64,6 @@ Global
 		{D24161F5-5B89-4EE5-94D3-660E9BC94CFA} = {FC5EA7BE-2E36-4CC7-8CE0-DD8D69F8CF85}
 		{BE392BCA-BDD7-43F6-BF11-9AC259F82CE6} = {C407EF64-970A-498F-989D-17BBEF83B36F}
 		{45D92BDD-5EA8-4B6C-A6F0-38C9DE139A32} = {BE392BCA-BDD7-43F6-BF11-9AC259F82CE6}
+		{68CBB995-C501-41BB-93B4-01A0CC618A89} = {CDC3A00A-52BD-4CDA-B7D7-1FCFC5DD1CEE}
 	EndGlobalSection
 EndGlobal
