+++ src/blueprints/Do.Blueprints.Service.Application/Do.Blueprints.Service.Application.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <AssemblyTitle>Do.Blueprints.Service.Application</AssemblyTitle>
     <Title>DO / Blueprints / Service</Title>
-    <Description>DO is an opinionated framework for .NET. This is a blueprint package that brings together all the features one needs for a simple service application.</Description>
+    <Description>DO is an opinionated framework for .NET. This is a blueprint package that brings together all the feature implementations one needs for a service application.</Description>
     <PackageId>do-blueprints-service-application</PackageId>
     <PackageTags>do;blueprint;service;application</PackageTags>
   </PropertyGroup>
@@ -26,6 +26,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Do.Blueprints.Service\Do.Blueprints.Service.csproj" />
     <ProjectReference Include="..\..\core\Do.Architecture\Do.Architecture.csproj" />
   </ItemGroup>
