+++ src/blueprints/Do.Blueprints.Service.Application/do-blueprints-service-application.targets
@@ -1,5 +1,7 @@
 <Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+
   <ItemGroup>
     <Using Include="Do" />
   </ItemGroup>
+
 </Project>
