+++ src/blueprints/Do.Blueprints.Service/Core/ISystem.cs
