+++ src/blueprints/Do.Blueprints.Service/Do.Blueprints.Service.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <AssemblyTitle>Do.Blueprints.Service</AssemblyTitle>
+    <Title>DO / Blueprints / Service</Title>
+    <Description>DO is an opinionated framework for .NET. This is a blueprint package that brings together all the feature abstractions one needs for a service domain.</Description>
+    <PackageId>do-blueprints-service</PackageId>
+    <PackageTags>do;blueprint;service</PackageTags>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Include="do-blueprints-service.targets" Pack="true" PackagePath="build" />
+  </ItemGroup>
+
+</Project>
